requirements.txt
selene==2.0.0rc4
pytest==7.4.3
allure-pytest==2.13.2
python-dotenv==1.0.0
